Added functionality to split det swagger doc into multiple definitions. Most (maybe too much) is still shared between the apis.
