+++ src/Umbraco.Cms.Api.Common/Attributes/MapToApiAttribute.cs
@@ -1,0 +1,9 @@
+namespace Umbraco.Cms.Api.Common.Attributes;
+
+[AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
+public class MapToApiAttribute : Attribute
+{
+    public MapToApiAttribute(string apiName) => ApiName = apiName;
+
+    public string ApiName { get; }
+}
