+++ src/Umbraco.Cms.Api.Common/Configuration/DefaultApiConfiguration.cs
@@ -1,0 +1,6 @@
+namespace Umbraco.Cms.Api.Common.Configuration;
+
+internal static class DefaultApiConfiguration
+{
+    public const string ApiName = "default";
+}
