+++ src/Umbraco.Cms.Api.Common/Umbraco.Cms.Api.Common.csproj
@@ -11,8 +11,11 @@
 <!--    <PackageReference Include="Microsoft.AspNetCore.App" Version="2.2.8" />-->
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning" Version="5.0.0" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning.ApiExplorer" Version="5.0.0" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Umbraco.Core\Umbraco.Core.csproj" />
+    <ProjectReference Include="..\Umbraco.Web.Common\Umbraco.Web.Common.csproj" />
+    
   </ItemGroup>
 </Project>
