+++ src/Umbraco.Cms.Api.Delivery/Configuration/DeliveryApiConfiguration.cs
@@ -1,0 +1,8 @@
+namespace Umbraco.Cms.Api.Delivery.Configuration;
+
+internal static class DeliveryApiConfiguration
+{
+    internal const string ApiTitle = "Umbraco Delivery API";
+
+    internal const string ApiName = "delivery";
+}
