+++ src/Umbraco.Cms.Api.Delivery/Controllers/DeliveryApiControllerBase.cs
@@ -1,5 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
+using Umbraco.Cms.Api.Common.Attributes;
 using Umbraco.Cms.Api.Common.Filters;
+using Umbraco.Cms.Api.Delivery.Configuration;
 using Umbraco.Cms.Api.Delivery.Filters;
 using Umbraco.Cms.Core;
 
@@ -10,6 +12,7 @@ namespace Umbraco.Cms.Api.Delivery.Controllers;
 [DeliveryApiAccess]
 [JsonOptionsName(Constants.JsonOptionsNames.DeliveryApi)]
 [LocalizeFromAcceptLanguageHeader]
+[MapToApi(DeliveryApiConfiguration.ApiName)]
 public abstract class DeliveryApiControllerBase : Controller
 {
 }
